Refactor: 기록 가능 기간 계산 위치 옮김
