+++ src/pages/main/Calendar/Calendar.jsx
@@ -4,14 +4,35 @@ import dayjs from 'dayjs';
 import { dayStyle } from './Calendar.style';
 import Masonry from '@mui/lab/Masonry';
 import { Box, Grid2 } from '@mui/material';
+import utc from 'dayjs/plugin/utc';
+import timezone from 'dayjs/plugin/timezone';
+import isSameOrAfter from 'dayjs/plugin/isSameOrAfter';
+
+dayjs.extend(isSameOrAfter);
+dayjs.extend(utc);
+dayjs.extend(timezone);
+
+export const isRecordable = (time) => {
+    const startOfPeriod = dayjs(`${dayjs(time).year()}-11-30`).startOf('day');
+    const endOfPeriod = startOfPeriod.add(31, 'day').startOf('day');
+    const today = dayjs(time).startOf('day');
+
+    if (
+        today.isBefore(endOfPeriod.add(1, 'day')) &&
+        today.isSameOrAfter(startOfPeriod)
+    )
+        return true;
+    return false;
+};
 
 const Calendar = ({ serverTime, hasWritten }) => {
     const today = dayjs(serverTime);
 
+    const recordable = isRecordable(serverTime);
+
     const lastDayWritten = hasWritten[31];
     return (
         <Box
-            Box
             sx={{
                 pr: ['0', '2.5rem'],
                 pl: ['0.25rem', '2.5626rem'],
@@ -40,6 +61,7 @@ const Calendar = ({ serverTime, hasWritten }) => {
                             date={index}
                             styleConfig={dayStyle[index]}
                             lastDayWritten={lastDayWritten}
+                            recordable={recordable}
                         />
                     );
                 })}
@@ -63,6 +85,7 @@ const Calendar = ({ serverTime, hasWritten }) => {
                         date={30}
                         styleConfig={dayStyle[30]}
                         lastDayWritten={lastDayWritten}
+                        recordable={recordable}
                     />
                 </Grid2>
                 <Grid2 size={6}>
@@ -72,6 +95,7 @@ const Calendar = ({ serverTime, hasWritten }) => {
                         date={31}
                         styleConfig={dayStyle[31]}
                         lastDayWritten={lastDayWritten}
+                        recordable={recordable}
                     />
                 </Grid2>
             </Grid2>
